Move SQL types specific tests to separate test class
